+++ dbms/src/Storages/KVStore/Read/ReadIndexWorker.cpp
@@ -12,168 +12,10 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include <Common/MemoryAllocTrace.h>
-#include <Common/Stopwatch.h>
-#include <Common/TiFlashMetrics.h>
-#include <Common/setThreadName.h>
-#include <Storages/KVStore/FFI/ProxyFFI.h>
-#include <Storages/KVStore/KVStore.h>
-#include <Storages/KVStore/Read/ReadIndexWorker.h>
-#include <fmt/chrono.h>
-
-#pragma GCC diagnostic push
-#ifdef __clang__
-#pragma GCC diagnostic ignored "-Wdeprecated-declarations"
-#endif
-// include to suppress warnings on NO_THREAD_SAFETY_ANALYSIS. clang can't work without this include, don't know why
-#include <grpcpp/security/credentials.h>
-#pragma GCC diagnostic pop
-
-#include <queue>
-
+#include <Storages/KVStore/Read/ReadIndexWorkerImpl.h>
 
 namespace DB
 {
-// #define ADD_TEST_DEBUG_LOG_FMT
-
-#ifdef ADD_TEST_DEBUG_LOG_FMT
-
-static Poco::Logger * global_logger_for_test = nullptr;
-static std::mutex global_logger_mutex;
-
-#define TEST_LOG_FMT(...) LOG_ERROR(global_logger_for_test, __VA_ARGS__)
-
-void F_TEST_LOG_FMT(const std::string & s)
-{
-    auto _ = std::lock_guard(global_logger_mutex);
-    std::cout << s << std::endl;
-}
-#else
-#define TEST_LOG_FMT(...)
-void F_TEST_LOG_FMT(const std::string &) {}
-#endif
-} // namespace DB
-
-namespace DB
-{
-AsyncNotifier::Status AsyncWaker::Notifier::blockedWaitUtil(const SteadyClock::time_point & time_point)
-{
-    // if flag from false to false, wait for notification.
-    // if flag from true to false, do nothing.
-    auto res = AsyncNotifier::Status::Normal;
-    if (!is_awake->exchange(false, std::memory_order_acq_rel))
-    {
-        {
-            auto lock = genUniqueLock();
-            if (!is_awake->load(std::memory_order_acquire))
-            {
-                if (condVar().wait_until(lock, time_point) == std::cv_status::timeout)
-                    res = AsyncNotifier::Status::Timeout;
-            }
-        }
-        is_awake->store(false, std::memory_order_release);
-    }
-    return res;
-}
-
-void AsyncWaker::Notifier::wake() NO_THREAD_SAFETY_ANALYSIS
-{
-    // if flag from false -> true, then wake up.
-    // if flag from true -> true, do nothing.
-    if (is_awake->load(std::memory_order_acquire))
-        return;
-    if (!is_awake->exchange(true, std::memory_order_acq_rel))
-    {
-        // wake up notifier
-        auto _ = genLockGuard();
-        condVar().notify_one();
-    }
-}
-
-void AsyncWaker::wake(RawVoidPtr notifier_)
-{
-    auto & notifier = *reinterpret_cast<AsyncNotifier *>(notifier_);
-    notifier.wake();
-}
-
-AsyncWaker::AsyncWaker(const TiFlashRaftProxyHelper & helper_)
-    : AsyncWaker(helper_, new AsyncWaker::Notifier{})
-{}
-
-AsyncWaker::AsyncWaker(const TiFlashRaftProxyHelper & helper_, AsyncNotifier * notifier_)
-    : inner(helper_.makeAsyncWaker(AsyncWaker::wake, GenRawCppPtr(notifier_, RawCppPtrTypeImpl::WakerNotifier)))
-    , notifier(*notifier_)
-{}
-
-AsyncNotifier::Status AsyncWaker::waitUtil(SteadyClock::time_point time_point)
-{
-    return notifier.blockedWaitUtil(time_point);
-}
-
-RawVoidPtr AsyncWaker::getRaw() const
-{
-    return inner.ptr;
-}
-
-class BlockedReadIndexHelperTrait
-{
-public:
-    explicit BlockedReadIndexHelperTrait(uint64_t timeout_ms_)
-        : time_point(SteadyClock::now() + std::chrono::milliseconds{timeout_ms_})
-    {}
-    virtual AsyncNotifier::Status blockedWaitUtil(SteadyClock::time_point) = 0;
-
-    // block current runtime and wait.
-    virtual AsyncNotifier::Status blockedWait()
-    {
-        // TODO: use async process if supported by framework
-        return blockedWaitUtil(time_point);
-    }
-    virtual ~BlockedReadIndexHelperTrait() = default;
-
-protected:
-    SteadyClock::time_point time_point;
-};
-
-class BlockedReadIndexHelper final : public BlockedReadIndexHelperTrait
-{
-public:
-    BlockedReadIndexHelper(uint64_t timeout_ms_, AsyncWaker & waker_)
-        : BlockedReadIndexHelperTrait(timeout_ms_)
-        , waker(waker_)
-    {}
-
-    const AsyncWaker & getWaker() const { return waker; }
-
-    AsyncNotifier::Status blockedWaitUtil(SteadyClock::time_point time_point) override
-    {
-        return waker.waitUtil(time_point);
-    }
-
-    ~BlockedReadIndexHelper() override = default;
-
-private:
-    AsyncWaker & waker;
-};
-
-class BlockedReadIndexHelperV3 final : public BlockedReadIndexHelperTrait
-{
-public:
-    BlockedReadIndexHelperV3(uint64_t timeout_ms_, AsyncWaker::Notifier & notifier_)
-        : BlockedReadIndexHelperTrait(timeout_ms_)
-        , notifier(notifier_)
-    {}
-
-    AsyncNotifier::Status blockedWaitUtil(SteadyClock::time_point time_point) override
-    {
-        return notifier.blockedWaitUtil(time_point);
-    }
-
-    ~BlockedReadIndexHelperV3() override = default;
-
-private:
-    AsyncWaker::Notifier & notifier;
-};
 
 BatchReadIndexRes TiFlashRaftProxyHelper::batchReadIndex(
     const std::vector<kvrpcpb::ReadIndexRequest> & req,
@@ -209,7 +51,8 @@ BatchReadIndexRes TiFlashRaftProxyHelper::batchReadIndex_v2(
         }
     }
 
-    { // block wait for all tasks are ready or timeout.
+    {
+        // Block wait for all tasks are ready or timeout.
         kvrpcpb::ReadIndexResponse tmp;
         while (!tasks.empty())
         {
@@ -227,7 +70,8 @@ BatchReadIndexRes TiFlashRaftProxyHelper::batchReadIndex_v2(
             }
         }
     }
-    { // if meet timeout, which means part of regions can not get response from leader, try to poll rest tasks
+    {
+        // If meets timeout, which means some of the regions can not get response from leader, try to poll rest tasks
         while (!tasks.empty())
         {
             auto & it = tasks.front();
@@ -286,57 +130,6 @@ bool TiFlashRaftProxyHelper::pollTimerTask(TimerTask & task, RawVoidPtr waker) c
     return fn_poll_timer_task(task.ptr, waker);
 }
 
-struct ReadIndexNotifyCtrl : MutexLockWrap
-{
-    using Data = std::deque<std::pair<RegionID, Timestamp>>;
-
-    bool empty() const NO_THREAD_SAFETY_ANALYSIS
-    {
-        auto _ = genLockGuard();
-        return data.empty();
-    }
-    void add(RegionID id, Timestamp ts) NO_THREAD_SAFETY_ANALYSIS
-    {
-        auto _ = genLockGuard();
-        data.emplace_back(id, ts);
-    }
-    Data popAll() NO_THREAD_SAFETY_ANALYSIS
-    {
-        auto _ = genLockGuard();
-        return std::move(data);
-    }
-
-    void wake() const { notifier->wake(); }
-
-    explicit ReadIndexNotifyCtrl(AsyncWaker::NotifierPtr notifier_)
-        : notifier(notifier_)
-    {}
-
-    Data data;
-    AsyncWaker::NotifierPtr notifier;
-};
-
-struct RegionReadIndexNotifier final : AsyncNotifier
-{
-    void wake() override
-    {
-        notify->add(region_id, ts);
-        notify->wake();
-    }
-    Status blockedWaitUtil(const SteadyClock::time_point &) override { return Status::Timeout; }
-
-    ~RegionReadIndexNotifier() override = default;
-
-    RegionReadIndexNotifier(RegionID region_id_, Timestamp ts_, const ReadIndexNotifyCtrlPtr & notify_)
-        : region_id(region_id_)
-        , ts(ts_)
-        , notify(notify_)
-    {}
-
-    RegionID region_id;
-    Timestamp ts;
-    ReadIndexNotifyCtrlPtr notify;
-};
 
 std::atomic<std::chrono::milliseconds> ReadIndexWorker::max_read_index_task_timeout
     = std::chrono::milliseconds{8 * 1000};
@@ -356,25 +149,6 @@ RegionNotifyMap::Data RegionNotifyMap::popAll() NO_THREAD_SAFETY_ANALYSIS
 {
     auto _ = genLockGuard();
     return std::move(data);
-}
-void ReadIndexDataNode::WaitingTasks::add(Timestamp ts, ReadIndexFuturePtr f) NO_THREAD_SAFETY_ANALYSIS
-{
-    auto _ = genLockGuard();
-    waiting_tasks.emplace_back(ts, std::move(f));
-}
-
-std::optional<ReadIndexDataNode::WaitingTasks::Data> ReadIndexDataNode::WaitingTasks::popAll() NO_THREAD_SAFETY_ANALYSIS
-{
-    auto _ = genLockGuard();
-    if (waiting_tasks.empty())
-        return {};
-    return std::move(waiting_tasks);
-}
-
-size_t ReadIndexDataNode::WaitingTasks::size() const NO_THREAD_SAFETY_ANALYSIS
-{
-    auto _ = genLockGuard();
-    return waiting_tasks.size();
 }
 
 void ReadIndexFuture::update(kvrpcpb::ReadIndexResponse resp) NO_THREAD_SAFETY_ANALYSIS
@@ -411,277 +185,6 @@ std::optional<kvrpcpb::ReadIndexResponse> ReadIndexFuture::poll(const std::share
     return resp;
 }
 
-void ReadIndexDataNode::ReadIndexElement::doTriggerCallbacks()
-{
-    TEST_LOG_FMT("start to triggerCallbacks for region_id={} callbacks size {}", region_id, callbacks.size());
-
-    if (resp.has_locked())
-    {
-        for (auto && cb : callbacks)
-        {
-            kvrpcpb::ReadIndexResponse res{};
-            if (cb->req.start_ts() == start_ts)
-                res = resp;
-            else
-                res.mutable_region_error()->mutable_epoch_not_match();
-            cb->update(std::move(res));
-        }
-    }
-    else
-    {
-        for (auto && cb : callbacks)
-        {
-            cb->update(resp);
-        }
-    }
-    callbacks.clear();
-}
-
-void ReadIndexDataNode::ReadIndexElement::doForceSet(const kvrpcpb::ReadIndexResponse & f)
-{
-    TEST_LOG_FMT("force set response {}, start-ts {}", f.ShortDebugString(), start_ts);
-
-    task_pair.reset();
-    resp = f;
-    doTriggerCallbacks();
-}
-
-void ReadIndexDataNode::ReadIndexElement::doPoll(
-    const TiFlashRaftProxyHelper & helper,
-    std::chrono::milliseconds timeout)
-{
-    bool can_trigger = false;
-    {
-        if (task_pair)
-        {
-            auto && [task, waker] = *task_pair;
-
-            auto * raw_waker = waker.getRaw();
-            bool clean_task = false;
-            if (helper.pollReadIndexTask(task, resp, raw_waker))
-            {
-                TEST_LOG_FMT(
-                    "poll ReadIndexElement success for region_id={}, resp {}",
-                    region_id,
-                    resp.ShortDebugString());
-
-                clean_task = true;
-            }
-            else if (SteadyClock::now() > timeout + start_time)
-            {
-                TEST_LOG_FMT("poll ReadIndexElement timeout for region_id={}", region_id);
-
-                clean_task = true;
-                resp.mutable_region_error()
-                    ->mutable_server_is_busy(); // set region_error `server_is_busy` for task timeout
-            }
-            else
-            {
-                TEST_LOG_FMT(
-                    "poll ReadIndexElement failed, region_id={} time_cost={} timeout={}",
-                    region_id,
-                    std::chrono::duration<double, std::milli>(SteadyClock::now() - start_time).count(),
-                    timeout);
-            }
-
-            if (clean_task)
-            {
-                can_trigger = true;
-                task_pair.reset();
-            }
-        }
-        else
-            can_trigger = true;
-    }
-    if (can_trigger)
-    {
-        doTriggerCallbacks();
-    }
-}
-
-void ReadIndexDataNode::doAddHistoryTasks(Timestamp ts, kvrpcpb::ReadIndexResponse && resp)
-{
-    for (auto it = running_tasks.begin(); it != running_tasks.end();)
-    {
-        if (it->first <= ts)
-        {
-            it->second.doForceSet(resp); // copy resp
-            it = running_tasks.erase(it);
-        }
-        else
-        {
-            break;
-        }
-    }
-    {
-        history_success_tasks.emplace(ts, std::move(resp)); // move resp
-    }
-}
-
-void ReadIndexDataNode::doConsume(const TiFlashRaftProxyHelper & helper, RunningTasks::iterator it)
-{
-    it->second.doPoll(helper, ReadIndexWorker::getMaxReadIndexTaskTimeout());
-    if (!it->second.task_pair)
-    {
-        auto start_ts = it->first;
-        auto resp = std::move(it->second.resp);
-
-        running_tasks.erase(it);
-
-        // If `start ts` is NOT initialized or response has `region error` | `lock`, just ignore
-        if (start_ts && !resp.has_locked() && !resp.has_region_error())
-        {
-            doAddHistoryTasks(start_ts, std::move(resp));
-        }
-    }
-}
-
-void ReadIndexDataNode::consume(const TiFlashRaftProxyHelper & helper, Timestamp ts) NO_THREAD_SAFETY_ANALYSIS
-{
-    auto _ = genLockGuard();
-
-    if (auto it = running_tasks.find(ts); it != running_tasks.end())
-    {
-        doConsume(helper, it);
-    }
-}
-
-const uint64_t MockStressTestCfg::RegionIdPrefix = 1ull << 40;
-bool MockStressTestCfg::enable = false;
-
-std::optional<ReadIndexTask> makeReadIndexTask(const TiFlashRaftProxyHelper & helper, kvrpcpb::ReadIndexRequest & req)
-{
-    if (likely(!MockStressTestCfg::enable))
-        return helper.makeReadIndexTask(req);
-    else
-    {
-        auto ori_id = req.context().region_id();
-        req.mutable_context()->set_region_id(
-            ori_id % MockStressTestCfg::RegionIdPrefix); // set region id to original one.
-        TEST_LOG_FMT("hacked ReadIndexTask to req {}", req.ShortDebugString());
-        auto res = helper.makeReadIndexTask(req);
-        req.mutable_context()->set_region_id(ori_id);
-        return res;
-    }
-}
-
-void ReadIndexDataNode::runOneRound(const TiFlashRaftProxyHelper & helper, const ReadIndexNotifyCtrlPtr & notify)
-    NO_THREAD_SAFETY_ANALYSIS
-{
-    auto opt_waiting_tasks = this->waiting_tasks.popAll();
-    if (!opt_waiting_tasks)
-        return;
-    auto & waiting_tasks = *opt_waiting_tasks;
-
-    auto _ = genLockGuard();
-
-    {
-        // Find the task with the maximum ts in all `waiting_tasks`.
-        Timestamp max_ts = 0;
-        ReadIndexFuturePtr max_ts_task = nullptr;
-        {
-            const ReadIndexFuturePtr * x = nullptr;
-            for (auto & e : waiting_tasks)
-            {
-                if (e.first >= max_ts)
-                {
-                    max_ts = e.first;
-                    x = &e.second;
-                }
-            }
-            max_ts_task = *x; // NOLINT
-        }
-
-        auto region_id = max_ts_task->req.context().region_id();
-
-        TEST_LOG_FMT(
-            "try to use max_ts {}, from request for region_id={}, waiting_tasks size {}, running_tasks {}",
-            max_ts,
-            region_id,
-            waiting_tasks.size(),
-            running_tasks.size());
-
-        // start-ts `0` will be used to only get the latest index, do not use history
-        if (history_success_tasks && history_success_tasks->first >= max_ts && max_ts)
-        {
-            TEST_LOG_FMT("find history_tasks resp {}", history_success_tasks->second.ShortDebugString());
-
-            for (const auto & e : waiting_tasks)
-            {
-                e.second->update(history_success_tasks->second);
-            }
-
-            cnt_use_history_tasks += waiting_tasks.size();
-        }
-        else
-        {
-            auto run_it = running_tasks.lower_bound(max_ts);
-            if (run_it == running_tasks.end())
-            {
-                TEST_LOG_FMT("no exist running_tasks for ts {}", max_ts);
-
-                if (auto t = makeReadIndexTask(helper, max_ts_task->req); t)
-                {
-                    TEST_LOG_FMT("successfully make ReadIndexTask for region_id={} ts {}", region_id, max_ts);
-                    AsyncWaker waker{helper, new RegionReadIndexNotifier(region_id, max_ts, notify)};
-                    run_it = running_tasks.try_emplace(max_ts, region_id, max_ts).first;
-                    run_it->second.task_pair.emplace(std::move(*t), std::move(waker));
-                }
-                else
-                {
-                    TEST_LOG_FMT("failed to make ReadIndexTask for region_id={} ts {}", region_id, max_ts);
-                    run_it = running_tasks.try_emplace(max_ts, region_id, max_ts).first;
-                    run_it->second.resp.mutable_region_error();
-                }
-            }
-
-            for (auto && e : waiting_tasks)
-            {
-                run_it->second.callbacks.emplace_back(std::move(e.second));
-            }
-
-            doConsume(helper, run_it);
-        }
-    }
-}
-
-ReadIndexDataNode::~ReadIndexDataNode() NO_THREAD_SAFETY_ANALYSIS
-{
-    auto _ = genLockGuard();
-
-    kvrpcpb::ReadIndexResponse resp;
-    resp.mutable_region_error()->mutable_region_not_found();
-
-    TEST_LOG_FMT(
-        "~ReadIndexDataNode region_id={}: waiting_tasks {}, running_tasks {} ",
-        region_id,
-        waiting_tasks.size(),
-        running_tasks.size());
-
-    if (auto waiting_tasks = this->waiting_tasks.popAll(); waiting_tasks)
-    {
-        for (const auto & e : *waiting_tasks)
-        {
-            e.second->update(resp);
-        }
-    }
-
-    for (auto & e : running_tasks)
-    {
-        e.second.doForceSet(resp);
-    }
-    running_tasks.clear();
-}
-
-ReadIndexFuturePtr ReadIndexDataNode::insertTask(const kvrpcpb::ReadIndexRequest & req)
-{
-    auto task = std::make_shared<ReadIndexFuture>();
-    task->req = req;
-
-    waiting_tasks.add(req.start_ts(), task);
-
-    return task;
-}
 
 ReadIndexDataNodePtr ReadIndexWorker::DataMap::upsertDataNode(RegionID region_id) const
 {
@@ -760,10 +263,6 @@ ReadIndexFuturePtr ReadIndexWorker::genReadIndexFuture(const kvrpcpb::ReadIndexR
     return res;
 }
 
-ReadIndexFuturePtr ReadIndexWorkerManager::genReadIndexFuture(const kvrpcpb::ReadIndexRequest & req)
-{
-    return getWorkerByRegion(req.context().region_id()).genReadIndexFuture(req);
-}
 
 void ReadIndexWorker::runOneRound(SteadyClock::duration min_dur)
 {
@@ -798,303 +297,14 @@ bool ReadIndexWorker::lastRunTimeout(SteadyClock::duration timeout) const
     return last_run_time.load(std::memory_order_relaxed) + timeout < SteadyClock::now();
 }
 
-ReadIndexWorker & ReadIndexWorkerManager::getWorkerByRegion(RegionID region_id)
-{
-    return *workers[region_id % workers.size()];
-}
-
-ReadIndexWorkerManager::ReadIndexWorkerManager(
-    const TiFlashRaftProxyHelper & proxy_helper_,
-    KVStore & kvstore_,
-    size_t workers_cnt,
-    ReadIndexWorkerManager::FnGetTickTime && fn_min_dur_handle_region,
-    size_t runner_cnt)
-    : proxy_helper(proxy_helper_)
-    , kvstore(kvstore_)
-    , logger(Logger::get("ReadIndexWorkers"))
-{
-    for (size_t i = 0; i < runner_cnt; ++i)
-        runners.emplace_back(std::make_unique<ReadIndexRunner>(
-            i,
-            runner_cnt,
-            kvstore,
-            workers,
-            logger,
-            fn_min_dur_handle_region,
-            std::make_shared<AsyncWaker::Notifier>()));
-
-    {
-        workers.reserve(workers_cnt);
-        for (size_t i = 0; i < workers_cnt; ++i)
-            workers.emplace_back(nullptr);
-
-        for (size_t rid = 0; rid < runner_cnt; ++rid)
-        {
-            for (size_t wid = rid; wid < workers_cnt; wid += runner_cnt)
-            {
-                workers[wid]
-                    = std::make_unique<ReadIndexWorker>(proxy_helper, kvstore, wid, runners[rid]->global_notifier);
-            }
-        }
-    }
-}
-
-void ReadIndexWorkerManager::wakeAll()
-{
-    for (auto & runner : runners)
-        runner->wake();
-}
-
-void ReadIndexWorkerManager::asyncRun()
-{
-    for (auto & runner : runners)
-        runner->asyncRun();
-}
-
-void ReadIndexWorkerManager::stop()
-{
-    for (auto & runner : runners)
-        runner->stop();
-}
-
-ReadIndexWorkerManager::~ReadIndexWorkerManager()
-{
-    stop();
-}
-
-void ReadIndexWorkerManager::runOneRoundAll(SteadyClock::duration min_dur)
-{
-    for (size_t id = 0; id < runners.size(); ++id)
-        runOneRound(min_dur, id);
-}
-
-void ReadIndexWorkerManager::runOneRound(SteadyClock::duration min_dur, size_t id)
-{
-    runners[id]->runOneRound(min_dur);
-}
-
-BatchReadIndexRes ReadIndexWorkerManager::batchReadIndex(
-    const std::vector<kvrpcpb::ReadIndexRequest> & reqs,
-    uint64_t timeout_ms)
-{
-    TEST_LOG_FMT("reqs size {}, timeout {}ms", reqs.size(), timeout_ms);
-
-    auto notifier = std::make_shared<AsyncWaker::Notifier>();
-    BlockedReadIndexHelperV3 helper{timeout_ms, *notifier};
-    std::queue<std::pair<RegionID, ReadIndexFuturePtr>> tasks;
-    BatchReadIndexRes resps;
-    resps.reserve(reqs.size());
-
-    for (const auto & req : reqs)
-    {
-        auto region_id = req.context().region_id();
-        auto & wk = this->getWorkerByRegion(region_id);
-        auto future = wk.genReadIndexFuture(req);
-        tasks.emplace(region_id, future);
-    }
-    this->wakeAll();
-
-    TEST_LOG_FMT("wake read_index_worker");
-
-    while (!tasks.empty())
-    {
-        auto & it = tasks.front();
-        if (auto res = it.second->poll(notifier); res)
-        {
-            resps.emplace_back(std::move(*res), it.first);
-            tasks.pop();
-        }
-        else
-        {
-            TEST_LOG_FMT("got resp {}, remain {}", resps.size(), tasks.size());
-            if (helper.blockedWait() == AsyncNotifier::Status::Timeout)
-            {
-                break;
-            }
-        }
-    }
-    { // if meet timeout, which means part of regions can not get response from leader, try to poll rest tasks
-        TEST_LOG_FMT("rest {}, poll rest tasks once", tasks.size());
-
-        while (!tasks.empty())
-        {
-            auto & it = tasks.front();
-            if (auto res = it.second->poll(); res)
-            {
-                resps.emplace_back(std::move(*res), it.first);
-            }
-            else
-            {
-                kvrpcpb::ReadIndexResponse tmp;
-                tmp.mutable_region_error()->mutable_region_not_found();
-                resps.emplace_back(std::move(tmp), it.first);
-            }
-            tasks.pop();
-        }
-    }
-    return resps;
-}
-
 void ReadIndexWorker::removeRegion(uint64_t region_id)
 {
     TEST_LOG_FMT("remove region_id={}", region_id);
     data_map.removeRegion(region_id);
 }
 
-BatchReadIndexRes KVStore::batchReadIndex(const std::vector<kvrpcpb::ReadIndexRequest> & reqs, uint64_t timeout_ms)
-    const
-{
-    assert(this->proxy_helper);
-    if (read_index_worker_manager)
-    {
-        return this->read_index_worker_manager->batchReadIndex(reqs, timeout_ms);
-    }
-    else
-    {
-        return proxy_helper->batchReadIndex_v1(reqs, timeout_ms);
-    }
-}
 
-std::unique_ptr<ReadIndexWorkerManager> ReadIndexWorkerManager::newReadIndexWorkerManager(
-    const TiFlashRaftProxyHelper & proxy_helper,
-    KVStore & kvstore,
-    size_t cap,
-    ReadIndexWorkerManager::FnGetTickTime && fn_min_dur_handle_region,
-    size_t runner_cnt)
-{
-#ifdef ADD_TEST_DEBUG_LOG_FMT
-    global_logger_for_test = &Poco::Logger::get("TestReadIndexWork");
-#endif
-    return std::make_unique<ReadIndexWorkerManager>(
-        proxy_helper,
-        kvstore,
-        cap,
-        std::move(fn_min_dur_handle_region),
-        runner_cnt);
-}
-
-void KVStore::initReadIndexWorkers(
-    ReadIndexWorkerManager::FnGetTickTime && fn_min_dur_handle_region,
-    size_t runner_cnt,
-    size_t worker_coefficient)
-{
-    if (!runner_cnt)
-    {
-        LOG_WARNING(log, "Run without read-index workers");
-        return;
-    }
-    auto worker_cnt = worker_coefficient * runner_cnt;
-    LOG_INFO(log, "Start to initialize read-index workers: worker count {}, runner count {}", worker_cnt, runner_cnt);
-    auto * ptr = ReadIndexWorkerManager::newReadIndexWorkerManager(
-                     *proxy_helper,
-                     *this,
-                     worker_cnt,
-                     std::move(fn_min_dur_handle_region),
-                     runner_cnt)
-                     .release();
-    std::atomic_thread_fence(std::memory_order_seq_cst);
-    read_index_worker_manager = ptr;
-}
-
-void KVStore::asyncRunReadIndexWorkers() const
-{
-    if (!read_index_worker_manager)
-        return;
-
-    assert(this->proxy_helper);
-    read_index_worker_manager->asyncRun();
-}
-
-void KVStore::stopReadIndexWorkers() const
-{
-    if (!read_index_worker_manager)
-        return;
-
-    assert(this->proxy_helper);
-    read_index_worker_manager->stop();
-}
-
-void KVStore::releaseReadIndexWorkers()
-{
-    if (read_index_worker_manager)
-    {
-        delete read_index_worker_manager;
-        read_index_worker_manager = nullptr;
-    }
-}
-
-void ReadIndexWorkerManager::ReadIndexRunner::wake() const
-{
-    global_notifier->wake();
-}
-
-void ReadIndexWorkerManager::ReadIndexRunner::stop()
-{
-    auto tmp = State::Running;
-    state.compare_exchange_strong(tmp, State::Stopping, std::memory_order_acq_rel);
-    global_notifier->wake();
-    if (work_thread)
-    {
-        work_thread->join();
-        work_thread.reset();
-        LOG_INFO(logger, "Thread of read-index runner {} has joined", id);
-    }
-    state.store(State::Terminated);
-}
-
-void ReadIndexWorkerManager::ReadIndexRunner::blockedWaitFor(std::chrono::milliseconds timeout) const
-{
-    global_notifier->blockedWaitFor(timeout);
-}
-
-void ReadIndexWorkerManager::ReadIndexRunner::runOneRound(SteadyClock::duration min_dur)
-{
-    for (size_t i = id; i < workers.size(); i += runner_cnt)
-        workers[i]->runOneRound(min_dur);
-}
-
-void ReadIndexWorkerManager::ReadIndexRunner::asyncRun()
-{
-    state = State::Running;
-    work_thread = std::make_unique<std::thread>([this]() {
-        std::string name = fmt::format("ReadIndexWkr-{}", id);
-        setThreadName(name.data());
-        auto [ptr_a, ptr_d] = getAllocDeallocPtr();
-        kvstore.reportThreadAllocInfo(name, ReportThreadAllocateInfoType::Reset, 0);
-        kvstore.reportThreadAllocInfo(name, ReportThreadAllocateInfoType::AllocPtr, reinterpret_cast<uint64_t>(ptr_a));
-        kvstore.reportThreadAllocInfo(
-            name,
-            ReportThreadAllocateInfoType::DeallocPtr,
-            reinterpret_cast<uint64_t>(ptr_d));
-        LOG_INFO(logger, "Start read-index runner {}", id);
-        while (true)
-        {
-            auto base_tick_timeout = fn_min_dur_handle_region();
-            blockedWaitFor(base_tick_timeout);
-            runOneRound(base_tick_timeout);
-            if (state.load(std::memory_order_acquire) != State::Running)
-                break;
-        }
-        kvstore.reportThreadAllocInfo(name, ReportThreadAllocateInfoType::Remove, 0);
-        LOG_INFO(logger, "Start to stop read-index runner {}", id);
-    });
-}
-
-ReadIndexWorkerManager::ReadIndexRunner::ReadIndexRunner(
-    size_t id_,
-    size_t runner_cnt_,
-    KVStore & kvstore_,
-    ReadIndexWorkers & workers_,
-    LoggerPtr logger_,
-    FnGetTickTime fn_min_dur_handle_region_,
-    AsyncWaker::NotifierPtr global_notifier_)
-    : id(id_)
-    , runner_cnt(runner_cnt_)
-    , kvstore(kvstore_)
-    , workers(workers_)
-    , logger(std::move(logger_))
-    , fn_min_dur_handle_region(std::move(fn_min_dur_handle_region_))
-    , global_notifier(std::move(global_notifier_))
-{}
+const uint64_t MockStressTestCfg::RegionIdPrefix = 1ull << 40;
+bool MockStressTestCfg::enable = false;
 
 } // namespace DB
